Anding functionalaty for loss augmentation on certen sections of the body - e.g. the arms and legs. This allows us to force the model to fokus on these parts instead on the fingeres

+++ DiffuseStyleGestureReproduced/debugger.py
@@ -71,9 +71,9 @@ class Debugger:
                     elif display == Show.IMAGE:
 
                         if tensor.dim() == 3:
-                            plt.imshow(tensor.permute(2,1,0)[:,:,0].cpu().detach().numpy())
+                            plt.imshow(tensor.permute(2,1,0)[:,:,0].cpu().to(torch.float32).detach().numpy())
                         elif tensor.dim() == 2:
-                            plt.imshow(tensor.permute(1, 0).cpu().detach().numpy())
+                            plt.imshow(tensor.permute(1, 0).cpu().to(torch.float32).detach().numpy())
                         
                         plt.title(tensor_name)
                         plt.show()
